# modules/api/api-documentation/api.yaml
openapi: 3.0.0
info:
  title: NDSP Core API
  description: API to gain access to core data
  version: 0.0.1-alpha
servers:
  - url: http://api.example.com/v1
    description: Optional server description, e.g. Main (production) server
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
  # /site:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
                  # type: string
  /site/{ods_code}:
    parameters:
      - name: ods_code
        in: path
        required: true
        schema:
          type: string
        description: The user ID
    get:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /transparency-site:
    get:
      summary: Returns an object with a count of transparencySites and the list of transparencySites
      description: >-
        In order to allow Users to see why their data could be accessed by a site or CCG. 
        This method will return a list of matching transparency sites.
      parameters:
        - name: provider_code
          in: query
          required: false
          schema: 
            type: string
        - name: provider_name
          in: query
          required: false
          schema: 
            type: string
        - name: start
          in: query
          required: false
          schema: 
            type: integer
        - name: count
          in: query
          required: false
          schema: 
            type: integer
      responses:
        '400': 
          description: Bad Request. No query terms where provided, exactly one must be provided
        '200': 
          description: An array of Direct Care API Sites who have signed the NDSA that match the provided search criteria
          content:
            application/json:
              schema: 
                type: object
                properties:
                  transparencySiteCount:
                    type: integer
                    format: int64
                    example: 27
                  transparencySites:
                    type: array
                    items: 
                      $ref: '#/components/schemas/transparencySite'
  /transparency-site/{provider_id}:
    get:
      summary: Returns a list of transparencySites
      description: >-
        In order to allow Users to see why their data could be accessed by a site or CCG. 
        This method will return a list of matching transparency sites.
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: 
            type: string
            format: uuid
      responses:
        '400': 
          description: Bad Request. Id is not valid
        '404': 
          description: Not Found. No matching site found for that Id 
        '200':   
          description: An array of Direct Care API Sites who have signed the NDSA that match the provided search criteria
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/transparencySite'
  # /ccg:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   put:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  # /care-setting:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  # /user:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #     tags:
  #       - User
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #     tags:
  #       - User
  #   put:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  # /product:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  # /supplier:
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   put:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  # /supplier/{id}/product:
  #   parameters:
  #     - name: id
  #       in: path
  #       required: true
  #       schema:
  #         type: string
  #       description: The user ID
  #   get:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  #   post:
  #     summary: Returns a list of users.
  #     description: Optional extended description in CommonMark or HTML.
  #     responses:
  #       '200':    # status code
  #         description: A JSON array of user names
  #         content:
  #           application/json:
  #             schema: 
  #               type: array
  #               items: 
  #                 type: string
  /acceptance:
    # get:
    #   summary: Returns a list of users.
    #   description: Optional extended description in CommonMark or HTML.
    #   responses:
    #     '200':    # status code
    #       description: A JSON array of user names
    #       content:
    #         application/json:
    #           schema: 
    #             type: array
    #             items: 
    #               type: string
    post:
      summary: Submit the acceptance data gather as part of an End User signup.
      description: >-
        In order to process the acceptance of the NDSA we need to capture the signatory details, 
        who they are signing on behalf of (organisation), the GP Connect interactions that they 
        are planning to use and reason for use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/acceptance'
      responses:
        '201':    # status code
          description: Created. The data has been persisted
        '400':
          description: Bad Request
    # put:
    #   summary: Returns a list of users.
    #   description: Optional extended description in CommonMark or HTML
    #   responses:
    #     '200':    # status code
    #       description: A JSON array of user names
    #       content:
    #         application/json:
    #           schema: 
    #             type: array
    #             items: 
    #               type: string

components:
  schemas:
    transparencySite:
      description: >-
        Transparency Sites are Direct Care API sites that have signed up to the NDSA. 
        The object contains details of the site and the use case for Data Access.
      type: object
      properties:
        id:
          type: string
          format: uuid
        odsCode:
          type: string
        name:
          type: string
        useCase:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        town:
          type: string
        county:
          type: string
        country:
          type: string
        postCode:
          type: string
        ccgIcbName:
          type: string
        ccgIcbOdsCode:
          type: string
        appointmentManagementEnabled:
          type: boolean
        accessRecordHTMLEnabled:
          type: boolean
        sendDocumentEnabled:
          type: boolean
        structuredRecordEnabled:
          type: boolean

    acceptance:
      description: >-
        Data captured as part of the NDSA Acceptance
      type: object
      properties:
        organisation:
          type: object
          properties:
            odsCode:
              type: string
            name:
              type: string
            addressLine1:
              type: string
            addressLine2:
              type: string
            town:
              type: string
            county:
              type: string
            country:
              type: string
            postCode:
              type: string
        useCase:
          type: string
        signatory:
          type: object
          properties:
            email: 
              type: string
            name: 
              type: string
            position: 
              type: string
        softwareSupplierName:
          type: string
        interactions:
          type: object
          properties:
            appointmentManagementEnabled:
              type: boolean
            accessRecordHTMLEnabled:
              type: boolean
            sendDocumentEnabled:
              type: boolean
            structuredRecordEnabled:
              type: boolean